+++ lib/main.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:subscript/other%20components/notificationservice.dart';
+import 'package:subscript/services/notificationservice.dart';
 import 'package:subscript/pages/homepage.dart';
 import 'dart:async';
-import 'package:subscript/other%20components/Subscription.dart';
+import 'package:subscript/services/subscription.dart';
 import 'package:subscript/pages/signinpage.dart';
 
 final lightTheme = ThemeData.light(useMaterial3: true);
